requirements.txt
google-generativeai
PyYAML
pydantic
pydantic-settings
tiktoken
pytest
python-dotenv
docker
